+++ arve/data/base_data.py
@@ -1,8 +1,16 @@
-from typing import Optional
+from arve import ARVE
+from typing import Optional, Union
 import numpy as np
 
 
-class add_spec:
+class BaseData:
+    """ARVE Data base-class."""
+
+    def __init__(self, arve: ARVE) -> None:
+        self.arve = arve
+        self.spec: dict[str, Union[np.ndarray, str, None]] = {}
+        self.vrad: dict[str, Union[np.ndarray, str, None]] = {}
+
     def add_spec(
         self,
         time: np.ndarray,
@@ -41,4 +49,30 @@ class add_spec:
             "time_unit": time_unit,
             "wave_unit": wave_unit,
             "flux_unit": flux_unit,
+        }
+
+    def add_vrad(
+        self,
+        time: np.ndarray,
+        vrad_val: np.ndarray,
+        vrad_err: Optional[np.ndarray] = None,
+        time_unit: Optional[str] = None,
+        vrad_unit: Optional[str] = None,
+    ) -> None:
+        """Add radial velocity data.
+
+        :param time: time values
+        :param vrad_val: radial velocity values
+        :param vrad_err: radial velocity errors, defaults to None
+        :param time_unit: time unit, defaults to None
+        :param vrad_unit: radial velocity unit, defaults to None
+        :return: None
+        """
+        # add dictionary with radial velocity data
+        self.vrad = {
+            "time": time,
+            "vrad_val": vrad_val,
+            "vrad_err": vrad_err,
+            "time_unit": time_unit,
+            "vrad_unit": vrad_unit,
         }
